+++ finagle-base-http/src/test/scala/com/twitter/finagle/http/util/HeaderKeyOrderingTest.scala
@@ -1,0 +1,22 @@
+package com.twitter.finagle.http.util
+
+import org.scalatest.FunSuite
+
+class HeaderKeyOrderingTest extends FunSuite {
+  test("keys are case insensitive") {
+    val key1 = "coffee"
+    val key2 = "COFFEE"
+    val key3 = "coFfEe"
+
+    // case difference ends up equal
+    assert(HeaderKeyOrdering.compare(key1, key2) == 0)
+    assert(HeaderKeyOrdering.compare(key2, key3) == 0)
+    assert(HeaderKeyOrdering.compare(key1, key3) == 0)
+  }
+
+  test("different keys are not equal") {
+    val key1 = "coffee"
+    val key2 = "covfefe"
+    assert(HeaderKeyOrdering.compare(key1, key2) != 0)
+  }
+}